refactor: onboarding components
