+++ mobile/src/app/(notes)/_layout.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { router, Stack } from 'expo-router';
 import { Ionicons } from '@expo/vector-icons';
 import { useTheme } from '@/hooks/useTheme';
-import Button from '@/components/reuseables/Button';
+import { HapticButton } from '@/components/layout';
 
 export default function NotesLayout() {
   const { icon } = useTheme();
@@ -14,9 +14,9 @@ export default function NotesLayout() {
           headerShown: true,
           headerTitle: '',
           headerLeft: () => (
-            <Button onPress={() => router.back()} hitSlop={10}>
+            <HapticButton onPress={() => router.back()} hitSlop={10}>
               <Ionicons name="chevron-back" size={24} color={icon} />
-            </Button>
+            </HapticButton>
           ),
         }}
       />
