+++ mobile/src/components/onboarding/index.ts
@@ -1,0 +1,7 @@
+export { default as NotesIllustration } from './NotesIllustration';
+export { default as OnboardingFooter } from './OnboardingFooter';
+export { default as OnboardingSlide } from './OnboardingSlide';
+export { default as ThemeSelector } from './ThemeSelector';
+export { default as WelcomeIllustration } from './WelcomeIllustration';
+
+export { onboardingSteps, type OnboardingStepProps } from '@/constants/onboardingSteps';
