+++ mobile/src/components/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as Button } from '../layout/HapticButton';
